+++ packages/plugin-packages/mashroom-storage/type-definitions/index.js
@@ -48,6 +48,10 @@ export interface MashroomStorageCollection<T: MashroomStorageRecord> {
      */
     updateOne(filter: MashroomStorageObjectFilter<T>, propertiesToUpdate: $Shape<MashroomStorageObject<T>>): Promise<MashroomStorageUpdateResult>;
     /**
+    * Update multiple entries
+    */
+    updateMany(filter: MashroomStorageObjectFilter<T>, propertiesToUpdate: $Shape<MashroomStorageObject<T>>): Promise<MashroomStorageUpdateResult>;
+    /**
      * Replace the first item that matches the given filter.
      */
     replaceOne(filter: MashroomStorageObjectFilter<T>, newItem: T): Promise<MashroomStorageUpdateResult>;
@@ -56,7 +60,7 @@ export interface MashroomStorageCollection<T: MashroomStorageRecord> {
      */
     deleteOne(filter: MashroomStorageObjectFilter<T>): Promise<MashroomStorageDeleteResult>;
     /**
-     * Delete all items that matches the given filter.
+     * Delete all items that match the given filter.
      */
     deleteMany(filter: MashroomStorageObjectFilter<T>): Promise<MashroomStorageDeleteResult>;
 }