+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Storage: Added a new method *updateMany* to update multiple entries at once
  * Portal: Allowed the Theme templates to access the full user, including *extraData*
  * Added health probes for the Remote App registry, so, the server will only be ready once the initial scan has
    been done (otherwise requests will hit instances with missing Apps).